+++ example/flutter_example/lib/computed.g.dart
@@ -13,7 +13,7 @@ const __myAgeMacroHash = 116800019;
 const myAge = 'MyAge: 32';
 
 const __randomNumberMacroHash = 340677358;
-const randomNumber = 262;
+const randomNumber = 352;
 
 const __myFavColorMacroHash = 906264747;
 const myFavColor = Color(4283215696);